Crate mock

Mock types and functions to generate GethData used for tests

Re-exports

pub use test_ctx::TestContext;

Modules

test_ctx

Mock types and functions to generate Test enviroments for ZKEVM tests

Structs

CORRECT_MOCK_TXS

Collection of correctly hashed and signed Transactions which can be used to test circuits or opcodes that have to check integrity of the Tx itself. Some of the parameters of the Tx are hardcoded such as nonce, value, gas_price etc…

MOCK_1_ETH

Mock 1 ETH

MOCK_ACCOUNTS

Mock accounts loaded with ETH to use for test cases.

MOCK_BASEFEE

Mock BASEFEE value

MOCK_CHAIN_ID

Mock chain ID value

MOCK_CODES

Mock EVM codes to use for test cases.

MOCK_COINBASE

Mock coinbase value

MOCK_DEPLOYED_CONTRACT_BYTECODE

Mock EVM bytecode for a deployed contract. PUSH1 0x20 PUSH1 0 PUSH1 0 CALLDATACOPY PUSH1 0x20 PUSH1 0 RETURN

MOCK_DIFFICULTY

Mock DIFFICULTY value

MOCK_GASLIMIT

Mock GASLIMIT value

MOCK_GASPRICE

Mock gasprice value

MOCK_WALLETS

Mock wallets used to generate correctly signed and hashed Transactions.

MockCallBytecodeParams

Holds the parameters for generating mock EVM bytecode for a contract call

MockTransaction

Mock structure which represents a Transaction and can be used for tests. It contains all the builder-pattern methods required to be able to specify any of it’s details.

Sha3CodeGen

Generate Sha3 opcode

Enums

AddrOrWallet

Constants

MOCK_BLOCK_GAS_LIMIT

Mock block gas limit

Functions

eth

Generate a [Word] which corresponds to a certain amount of ETH.

generate_mock_call_bytecode

Generate mock EVM bytecode that performs a contract call

gwei

Express an amount of ETH in GWei.

rand_bytes

Generate random bytes for the specified size.